test(tempo): isolate multisig chain config tests
