+++ src/tempo/chainConfig.test.ts
@@ -1,4 +1,3 @@
-import { MultisigConfig } from 'ox/tempo'
 import { describe, expect, test, vi } from 'vitest'
 import { accounts, feeToken, getClient } from '~test/tempo/config.js'
 import { generatePrivateKey } from '../accounts/generatePrivateKey.js'
@@ -323,60 +322,6 @@ describe('prepareTransactionRequest', () => {
     })
     const request = await prepareTransactionRequest(clientWithFeeToken, {})
     expect(request.feeToken).toBe(feeToken)
-  })
-
-  test('behavior: multisigSignatureCount left for node inference', async () => {
-    const config = MultisigConfig.from({
-      threshold: 2,
-      owners: [
-        { owner: accounts[1].address, weight: 1 },
-        { owner: accounts[2].address, weight: 1 },
-        { owner: accounts[3].address, weight: 1 },
-      ],
-    })
-
-    const request = await prepareTransactionRequest(client, {
-      multisig: config,
-      multisigVersion: 0n,
-      parameters: ['chainId'],
-    })
-
-    expect(request.multisigSignatureCount).toBeUndefined()
-  })
-
-  test('behavior: explicit multisigSignatureCount is preserved', async () => {
-    const config = MultisigConfig.from({
-      threshold: 2,
-      owners: [
-        { owner: accounts[1].address, weight: 1 },
-        { owner: accounts[2].address, weight: 1 },
-        { owner: accounts[3].address, weight: 1 },
-      ],
-    })
-
-    const request = await prepareTransactionRequest(client, {
-      multisig: config,
-      multisigSignatureCount: 3,
-      multisigVersion: 0n,
-      parameters: ['chainId'],
-    })
-
-    expect(request.multisigSignatureCount).toBe(3)
-  })
-
-  test('behavior: explicit multisigVersion is preserved', async () => {
-    const config = MultisigConfig.from({
-      threshold: 1,
-      owners: [{ owner: accounts[1].address, weight: 1 }],
-    })
-
-    const request = await prepareTransactionRequest(client, {
-      multisig: config,
-      multisigVersion: 2n,
-      parameters: ['chainId'],
-    })
-
-    expect(request.multisigVersion).toBe(2n)
   })
 
   test('behavior: keyAuthorizationManager attaches pending key authorization', async () => {
